@@ -1,27 +1,27 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-buildscript { 
-    ext.kotlin_version = '1.2.50'
-
-    
+buildscript {
+    ext.kotlin_version = '1.7.20'
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.4'
-        classpath "io.realm:realm-gradle-plugin:5.1.0"
-        
+        classpath 'com.android.tools.build:gradle:7.3.1'
+        classpath "io.realm:realm-gradle-plugin:10.11.1"
+
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
-    classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version" }
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+    }
 }
 
 allprojects {
     repositories {
         google()
-        jcenter()
+        maven { url 'https://jitpack.io' }
+        mavenCentral()
     }
 }
 
