@@ -1,8 +1,7 @@
 package fr.vpm.mypix;
 
-import android.app.Activity;
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.Collections;
