@@ -1,7 +1,7 @@
 package fr.vpm.mypix.album;
 
 import android.net.Uri;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import java.util.ArrayList;
 import java.util.List;
