@@ -1,6 +1,6 @@
 package fr.vpm.mypix.album;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public interface Picture {
 
