@@ -1,6 +1,6 @@
 package fr.vpm.mypix.flickr;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import fr.vpm.mypix.album.Album;
