@@ -1,7 +1,7 @@
 package fr.vpm.mypix.flickr;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import java.util.ArrayList;
